% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.models.r
\name{compare.models}
\alias{compare.models}
\title{Statistical comparison of pairs of either \code{esth} or \code{ggcline} objects with different pooling or fixing of parameters, using the 
  Bayesian widely applicable information criterion (waic).}
\usage{
compare.models(ggcline.object1 = NULL, ggcline.object2 = NULL,
  ggcline.pooled = FALSE, esth.object1 = NULL, esth.object2 = NULL)
}
\arguments{
\item{ggcline.object1}{Name of a \code{ggcline} object.}

\item{ggcline.object2}{Name of a \code{ggcline} object.}

\item{ggcline.pooled}{Logical. If \code{FALSE}, each row of the \code{ggcline$gc} objects is compared separately. If 
\code{TRUE}, a single comparison is made across all rows (to be used when one or both parameters is pooled across 
all samples in at least one of the two models being compared).}

\item{esth.object1}{Name of an \code{esth} object.}

\item{esth.object2}{Name of an \code{esth} object.}
}
\value{
A \code{data.table} and \code{data.frame} with the name of the test subject plus following fields:
  \item{npar1}{number of parameters for the first model.}
  \item{lppd1}{Sum of the mean posterior likelihoods for the individual data points for the first model.}
  \item{waic1}{widely applicable information criterion for the first model.}
  \item{se1}{standard error of waic1.}
  \item{npar2}{number of parameters for the second model.}
  \item{lppd2}{Sum of the mean posterior likelihoods for the individual data points for the second model.}
  \item{waic2}{widely applicable information criterion for the second model.}
  \item{se2}{standard error of waic2.}
  \item{diff}{difference in waic between the first and second models. A positive value favours the first model.}
}
\description{
Statistical comparison of pairs of either \code{esth} or \code{ggcline} objects with different pooling or fixing of parameters, using the 
  Bayesian widely applicable information criterion (waic).
}
\details{
\code{compare.models} uses the actual number of parameters for each model rather than an estimate,
  which would be applicable in more complex hierarchichal models.
}
\examples{

\dontrun{
#A comparison of individual-level hybrid index estimates with population-level hybrid index estimates, to see if there is 
#evidence of individuals within populations differing in hybrid index:
esthcomp=compare.models(hest.object1=hindlabel,
hest.object2=hpopid)
#A comparison of genomic cline estimates versus likelihoods of genomic clines with the two parameters fixed to null values (an 
#alternative way of testing significance to the p values presented in \\code{ggcline}):
gclinecomp=compare.models(ggcline.object1=gc1,
ggcline.object2=gcAllFixed)
}
}
\author{
Richard Ian Bailey \email{richardianbailey@gmail.com}
}
