% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.prep.r
\name{data.prep}
\alias{data.prep}
\title{Prepare data for hybrid index and genomic cline estimation.}
\usage{
data.prep(data, loci, alleles, S0, S1, precols, INDLABEL.name = "INDLABEL",
  POPID.name = "POPID", max.S.MAF = 0.5, min.diff = 0,
  min.allele.copies.S0 = 0, min.allele.copies.S1 = 0,
  return.genotype.table = FALSE, return.locus.table = FALSE,
  marker.info.file = NULL)
}
\arguments{
\item{data}{A \code{data} object produced by \code{read.data}, or a custom \code{data.table} in the same format.}

\item{loci}{A \code{loci} object produced by \code{read.data}, or a custom character vector with locus names.}

\item{alleles}{An \code{alleles} object produced by \code{read.data}, or a custom \code{data.table} in the same format.}

\item{S0}{Character vector. Names, taken from the designated \code{POPID} column in the \code{data} object, 
indicating which populations represent the first parental reference set (\dQuote{Source 0}). Allele frequencies 
in this set represent a hybrid index of 0.}

\item{S1}{Character vector. Names, taken from the designated \code{POPID} column in the \code{data} object, 
indicating which populations represent the second parental reference set (\dQuote{Source 1}). Allele frequencies 
in this set represent a hybrid index of 1.}

\item{precols}{A \code{precols} object produced by \code{read.data}, or a custom character vector with 
names of all pre-marker columns in the \code{data}.}

\item{INDLABEL.name}{Character string. The name of the column in \code{data} designated as \code{INDLABEL}. 
Default is \dQuote{INDLABEL}.}

\item{POPID.name}{Character string. The name of the column in \code{data} designated as \code{POPID}. 
Default is \dQuote{POPID}.}

\item{max.S.MAF}{Numeric locus filtering option. Removes loci for which the smaller of the two parental reference set minor allele 
frequencies is greater than this value. Included because only one parental reference minor allele frequency needs to be close to 
zero for the locus to be informative. Default is \code{0.5} (no filtering).}

\item{min.diff}{Numeric locus filtering option. Removes loci for which the difference in allele frequency between 
parental reference sets is less than this value. Default is \code{0} (no filtering).}

\item{min.allele.copies.S0}{Locus filtering option based on sample size. Removes loci for which the 
number of allele copies in the \code{S0} 
parental reference set, excluding missing data, is less than this value. Default is \code{0} (no filtering).}

\item{min.allele.copies.S1}{Locus filtering option based on sample size. Removes loci for which the 
number of allele copies in the \code{S1} 
parental reference set, excluding missing data, is less than this value. Default is \code{0} (no filtering).}

\item{return.genotype.table}{Logical. Whether to create a genotype table with individual genotypes for each locus scored from 
\code{0} to \code{PLOIDY}, based on their number of copies of the designated \code{S1} allele (the allele with relatively higher 
frequency in the \code{S1} parental reference set). Default is \code{FALSE}.}

\item{return.locus.table}{Logical. Whether to create a shortened version of the output table, with one row 
per locus and only containing locus-level data (such as parental allele frequencies). Default is \code{FALSE}.}

\item{marker.info.file}{Character string. The name of an optional file of additional marker information, to be read in 
and joined to the primary output table (and to the locus table if \code{return.locus.table} set to \code{TRUE}). 
The file should have one row per marker with marker names in a column named "locus".
May be useful for downstream statistical comparison of hybrid index or genomic cline estimates among groups of markers. 
Default is \code{NULL}.}
}
\value{
list with one mandatory component, \code{data.prep}, a \code{data.table} and \code{data.frame} long-form 
  version of the imported data table
  with additional fields listed below. The two optional components are: (1) \code{geno.data}, a \code{data.table} and \code{data.frame} 
  containing genotypes for each locus scored from 
  \code{0} to \code{PLOIDY}, based on their number of copies of the designated \code{S1} allele (the allele with relatively higher 
  frequency in the \code{S1} parental reference set); (2) \code{locus.data}, a shortened version of the output table, with one row 
  per locus and containing the same locus-level data as \code{data.prep}.

  \code{data.prep} contains the following additional fields:
  \item{Source}{whether the sample is from source 0 (\sQuote{S0}), source 1 (\sQuote{S1}) or is a test individual 
     (\sQuote{TEST}).}
  \item{S0_allele}{name of the allele with relatively higher frequency in \code{S0}.}
  \item{S1_allele}{name of the allele with relatively higher frequency in \code{S1}.}
  \item{S0.prop_1}{frequency (proportion) of the \code{S1_allele} in \code{S0}.}
  \item{S1.prop_1}{frequency (proportion) of the \code{S1_allele} in \code{S1}.}
  \item{S0.N_1}{number of non-missing copies of the \code{S1_allele} in \code{S0}.}
  \item{S1.N_1}{number of non-missing copies of the \code{S1_allele} in \code{S1}.}
  \item{S0.prop_0}{frequency (proportion) of the \code{S0_allele} in \code{S0}.}
  \item{S1.prop_0}{frequency (proportion) of the \code{S0_allele} in \code{S1}.}
  \item{N_allele_copies.S0}{total number of non-missing allele copies in \code{S0}.}
  \item{N_allele_copies.S1}{total number of non-missing allele copies in \code{S1}.}
  \item{Source.afdiff}{difference in allele frequency between \code{S0} and \code{S1}.}
  \item{S0.MAF}{\code{S0} minor allele frequency.}
  \item{S1.MAF}{\code{S1} minor allele frequency.}
  \item{min.MAF}{smaller of \code{S0.MAF} and \code{S1.MAF}.}
  \item{Source_allele}{numeric. Whether the allele is the \code{S0_allele} (\sQuote{0}) or \code{S1_allele} (\sQuote{1}).}
}
\description{
Prepare data for hybrid index and genomic cline estimation.
}
\details{
The primary output of \code{data.prep} is a long-form \code{data.table} with 1 row for each allele copy in the 
  data set, to be used in downstream hybrid index and genomic cline estimation. It also optionally produces (1) a locus table,
  with one row per locus and containing the same locus data as in the long-form table (see \sQuote{Value}), and (2) a 
  genotype table, with genotypes 
  scored according to the number of alleles from parental reference set \code{S1}. This may be useful for example in estimating 
  parent-specific linkage disequilibria. \code{data.prep} uses objects produced by \code{read.data} as input, or custom 
  objects in the same format.

The optional \code{marker.info.file} should be a file with one row per marker with marker names in a column named \dQuote{locus} and any number of 
  columns, each for a variable of
  interest for grouping markers, such as chromosome, map location, gene function, SNP type (non-synonymous, synonymous, 
  intronic etc.). These variables 
  can then be used downstream for statistical comparison of hybrid index or clines among groups of loci.
}
\examples{

\dontrun{
#Using objects produced by \\code{read.data}:
prepdata=data.prep(data=dat$data,
loci=dat$loci,
alleles=dat$alleles,
S0=c("pop1","pop2"),  #Two populations in the 'POPID' column chosen as parental reference set 'S0'#
S1="pop3",            #One population in the 'POPID' column chosen as parental reference set 'S1'#
precols=dat$precols,
max.S.MAF = 0.1,      #Filtering by parental minor allele frequency only; other filtering options below are same as default#
min.diff=0,
min.allele.copies.S0=0,
min.allele.copies.S1=0,
return.genotype.table=T,
return.locus.table=T)
}
}
\author{
Richard Ian Bailey \email{richardianbailey@gmail.com}
}
