% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esth.r
\name{esth}
\alias{esth}
\title{Bayesian hybrid index estimation.}
\usage{
esth(data.prep.object, read.data.precols, test.subject = "INDLABEL",
  test.subject.compare = "INDLABEL", include.Source = TRUE,
  return.likmeans = FALSE, fix.subject = FALSE, fix.value = FALSE,
  plot.ind = NULL, plot.col = NULL, nitt, burnin, init.var = 0.002,
  prior = c(0.5, 0.5), print.k = 50)
}
\arguments{
\item{data.prep.object}{Name of the \code{data.prep} object produced by the \code{data.prep} function.}

\item{read.data.precols}{A \code{precols} object produced by \code{read.data}, or a custom character vector with 
names of all pre-marker columns in the \code{data}.}

\item{test.subject}{Character string. Name of the field identifying subjects for hybrid index estimation. Default \dQuote{INDLABEL}.}

\item{test.subject.compare}{Character string. Name of a different field identifying subjects for 
hybrid index estimation, to be run through \code{esth} separately and compared with the current run 
downstream using \code{compare.models}.}

\item{include.Source}{Logical. Whether to estimate hybrid indices for test subjects declared as parental reference. Default is \code{TRUE}.}

\item{return.likmeans}{Logical. Whether to return a table of likelihood calculations. Required for \code{compare.models}.}

\item{fix.subject}{Logical default, otherwise a character vector. List of test subjects (normally individual references 
in the \dQuote{INDLABEL} field) for which you wish to fix rather than estimate the hybrid index. Potentially useful for model 
comparison with \code{compare.models}.}

\item{fix.value}{Logical default, otherwise a numeric vector of length 1 or the same length as \code{fix.subject}. 
The fixed hybrid index value(s) for the test subjects listed in \code{fix.subject}.}

\item{plot.ind}{Character vector. List of up to 6 test subjects for which you wish to plot the posterior distribution values in real time. 
Default \code{NULL}.}

\item{plot.col}{Character vector. List of up to 6 colours to plot the \code{plot.ind} test subjects. Default \code{NULL}.}

\item{nitt}{Numeric. The total number of MCMC iterations including burnin. At least 6000 is recommended.}

\item{burnin}{Numeric. The number of burnin MCMC iterations. At least 3000 is recommended.}

\item{init.var}{Numeric. Starting value for the variance of the parameter proposal distribution. For data sets with 
very large numbers of loci it may be useful to reduce this number as uncertainty in hybrid index will be very small. Default 
is 0.002, which works well with tested data sets.}

\item{prior}{Numeric vector. Values of \code{shape1} and \code{shape2} parameters for the beta-distributed prior. 
Default is \code{shape1=shape2=0.5} (Jeffrey's prior).}

\item{print.k}{Numeric. The iteration is printed to screen on multiples of this number. Default is \code{50}.}
}
\value{
list containing \code{hi}, \code{test.subject} and \code{likmeans} (optional). \code{hi} is a \code{data.table} 
  and \code{data.frame} containing all pre-marker columns that do not have finer resolution 
  than the declared \code{test.subject}, 
  as well as the following fields:
  \item{Source}{whether the sample is from source 0 ('S0'), source 1 ('S1') or is a test individual ('TEST').}
  \item{h_posterior_mode}{the hybrid index estimate (mode of the beta-distributed posterior).}
  \item{h_cred_int_lower}{lower 95 percent credible interval (2.5 percent quantile of the posterior beta distribution).}
  \item{h_cred_int_upper}{upper 95 percent credible interval (97.5 percent quantile of the posterior beta distribution).}
  \item{beta_mean}{mean of the posterior beta distribution (included for convenience).}
  \item{beta_var}{variance of the posterior beta distribution (included for convenience).}
  \item{beta_shape1}{beta shape1 parameter estimate for the posterior beta distribution (included for convenience).}
  \item{beta_shape2}{beta shape2 parameter estimate for the posterior beta distribution (included for convenience).}
  \item{npar_compare}{see compare.models.}
  \item{npar}{number of parameters. see compare.models.}
}
\description{
Bayesian hybrid index estimation.
}
\details{
\code{esth} estimates hybrid index using the likelihood formulae of Buerkle (2005), with the addition of a
  prior. The default is Jeffrey's prior (beta distribution with \code{shape1=shape2=0.5}), which testing suggests 
  is an improvement over a uniform prior (\code{shape1=shape2=1}).

\code{esth} excludes any pre-marker columns from the output with finer resolution than the declared \code{test.subject}, 
  and hence has the same number of rows as the number of unique \code{test.subject} values.

Set \code{return.likmeans=TRUE} if you intend to carry out model comparison.
}
\examples{

\dontrun{
hindlabel= esth(data.prep.object = prepdata$data.prep,
read.data.precols = dat$precols,
test.subject="INDLABEL",
test.subject.compare="POPID",
include.Source = TRUE,
return.likmeans = TRUE,
plot.ind = c("indref1","indref2","indref3","indref4","indref5","indref6"),
plot.col = c("blue","green","cyan","purple","magenta","red"),
nitt=15000,burnin=5000)
}
}
\references{
Buerkle, C. A. (2005). Maximum likelihood estimation of a hybrid index based on molecular markers. 
  Molecular Ecology Notes, 5(3), 684-687.
}
\author{
Richard Ian Bailey \email{richardianbailey@gmail.com}
}
