% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.data.r
\name{read.data}
\alias{read.data}
\title{Load data stored in the format of a \sQuote{structure} file or similar data table.}
\usage{
read.data(file, mainparams = NULL, extracol.names = NULL,
  precol.headers = 1, nprecol, markername.dup = 0, EXTRACOL = 0,
  INDLABEL = 1, LOCDATA = 0, MAPDISTANCE = 0, MARKERNAME = 1,
  MISSINGVAL, NUMINDS = 0, NUMLOCI = 0, ONEROW = 0, PHASED = 0,
  PHENOTYPE = 0, PLOIDY = 2, POPFLAG = 0, POPID = 1,
  RECESSIVEALLELE = 0)
}
\arguments{
\item{file}{Character string. The name of the data file to read in.}

\item{mainparams}{Character string. The name of an associated \code{mainparams} 
file to read in. Optional as \code{mainparams} info can be entered into the 
function call directly (see below), or may be unnecessary (see \sQuote{Details}).}

\item{extracol.names}{Character string or vector. Names of extra (i.e. none of those specifically named here) 
non-marker columns in the data file. Optional; used 
when the data file does not contain pre-marker \sQuote{extracol} headers and you 
wish to add them on input. Default is \code{NULL}.}

\item{precol.headers}{Numeric. Presence (\code{1}) or absence (\code{0}) of headers for 
pre-marker columns in the header row of the data file. Default is \code{1}.}

\item{nprecol}{Numeric. Number of pre-marker columns in the data file. No 
default and must always be entered.}

\item{markername.dup}{Numeric. Whether each marker name appears twice in 
the data file header row. Default is \code{0} (FALSE), alternative is \code{1} (TRUE).}

\item{EXTRACOL}{Numeric. Presence (\code{1}) or absence (\code{0}) of extra (i.e. none of those 
specifically named here) non-marker columns in the data file. Not needed if this info is 
uploaded in a \code{mainparams} file. For compatibility with structure \code{mainparams} info. Default is \code{0}.}

\item{INDLABEL}{Numeric. Presence (\code{1}) or absence (\code{0}) of a column of 
individual references in the data file. Not needed if the info is uploaded in a \code{mainparams} file. Default is \code{1}, and this column 
must always be present. See \sQuote{Details}.}

\item{LOCDATA}{Numeric. Presence (\code{1}) or absence (\code{0}) of a \code{LOCDATA} column in 
the data file. For compatibility with sQuote{structure}. Not needed if the info is uploaded in a \code{mainparams} file. Default is \code{0}.}

\item{MAPDISTANCE}{Numeric. Presence (\code{1}) or absence (\code{0}) of a \code{MAPDISTANCE} 
row in the data file, which will be removed if present. Not needed if the info is uploaded in a \code{mainparams} file. Default is \code{0}.}

\item{MARKERNAME}{Numeric. Presence (\code{1}) or absence (\code{0}) of a header row 
containing marker names in the data file. Not needed if the info is uploaded in a \code{mainparams} file. Default is \code{1}.}

\item{MISSINGVAL}{The identifier for missing values. No default. Not needed if the info is uploaded in a \code{mainparams} file.}

\item{NUMINDS}{Numeric. Number of individuals in the data set. Not used 
in downstream functions and included for user's convenience. Default 
set to \code{0}.}

\item{NUMLOCI}{Numeric. Number of loci in the data set. Not needed if the info is uploaded in a \code{mainparams} file. Default is \code{0}, but 
the correct number must be entered for most input file formats (see 
\sQuote{Details}).}

\item{ONEROW}{Numeric. Whether (\code{1}) or not (\code{0}) there is a single data row 
per \code{INDLABEL}. If the value is \code{1}, the number of columns per marker = 
\code{PLOIDY}. If \code{ONEROW = 0}, there is one column per marker and the number 
of rows per \code{INDLABEL} = \code{PLOIDY}. Not needed if the info is uploaded in a \code{mainparams} file. Default is \code{0}.}

\item{PHASED}{Numeric. Presence (\code{1}) or absence (\code{0}) of phasing data rows 
in a \sQuote{structure} input file. These rows will be removed if present. Not needed if the info is uploaded in a \code{mainparams} file. 
Default is \code{0}.}

\item{PHENOTYPE}{Numeric. Presence (\code{1}) or absence (\code{0}) of a \code{PHENOTYPE} 
column in the data file. For compatibility with \sQuote{structure}. Not needed if the info is uploaded in a \code{mainparams} file. Default 
is \code{0}.}

\item{PLOIDY}{Numeric. Maximum ploidy among the markers in the data file. Not needed if the info is uploaded in a \code{mainparams} file. 
Default is \code{2}.}

\item{POPFLAG}{Numeric. \code{0} or \code{1}, for compatibility with structure. Not 
needed and will be removed in future. Default \code{0}.}

\item{POPID}{Numeric. Presence (\code{1}) or absence (\code{0}) of a column of population 
identifiers in the data file (which should always be present, see \sQuote{Details}). Not needed if the info is uploaded in a \code{mainparams} file. 
Default is \code{1}.}

\item{RECESSIVEALLELE}{Numeric. \code{0} or \code{1}, for compatibility with \sQuote{structure}. Not 
needed and will be removed in future. Default \code{0}.}
}
\value{
\code{read.data} returns a list with one component containing the loaded data in \code{ONEROW = 0} format, in the 
  form of a \code{data.table} and \code{data.frame}, and other components used in downstream functions or otherwise potentially 
  useful to the user.

  The list contains the following components:
  \item{mainparams}{A \code{data.table} and \code{data.frame} with the inputted \code{mainparams} information (or default values if 
  no information provided).}
  \item{nprecols}{Numeric. The number of non-marker columns to the left of the marker columns in the imported data set.}
  \item{precols}{Character vector. The names of non-marker columns to the left of the marker columns in the imported data set.}
  \item{data}{A \code{data.table} and \code{data.frame}. The imported data.}
  \item{alleles}{A \code{data.table} and \code{data.frame}. The names of the two alleles at each locus.}
  \item{loci}{Character vector. The names of all loci in the imported data set.}
}
\description{
Load data stored in the format of a \sQuote{structure} file or similar data table.
}
\details{
\code{read.data} is designed to be compatible with data input files 
  for the software \href{http://web.stanford.edu/group/pritchardlab/structure.html}{structure}, including those produced by 
  \href{https://www.cog-genomics.org/plink2/}{PLINK}. However, the simplest file format to read in is
  a data table with a complete header row, one column per marker and one row per allele copy (equivalent to \code{ONEROW
  = 0}). With this format the only required fields
  are \code{file}, \code{nprecol} and \code{MISSINGVAL}. Furthermore, allele identities are 
  read in as character strings and therefore each allele can be given any 
  name or number, as long as there are only two unique alleles per locus. 
  As with \href{http://web.stanford.edu/group/pritchardlab/structure.html}{structure}, all non-marker 
  columns should be to the left of the marker columns in the data file.

The data input file is required to have at least two columns prior to the 
  marker data columns. The first is referred to as \sQuote{INDLABEL} (but can be 
  given a different column header) and represents the finest resolution of 
  identification that may be required for the data set (usually a unique 
  individual reference, but could for example be a population-level 
  reference for poolseq data). The second is referred to as \sQuote{POPID} and 
  indicates the population from which the individual was sampled. The main 
  purpose of this column is to declare, downstream, which \code{POPID} values 
  represent the parental reference samples, and it can also be useful for 
  plotting hybrid index estimates.

Any ploidy is allowed, but the declared \code{PLOIDY} must be the same for all markers. So for example if 
  data are haplodiploid, haploid markers should either be present as 
  diploid homozygotes, or (preferably from a statistical viewpoint) the 
  second allele for each \code{INDLABEL} and marker should be declared as missing data.

If an associated \code{mainparams} file is read in, it should contain two columns 
  and no header row, with the first column holding the field names and the 
  second the field values. The only mandatory fields are  
  (and must have one of these synonyms): \code{NUMLOCI}, \code{PLOIDY}, 
  \code{MISSINGVAL}/\code{MISSING}, \code{ONEROW}/\code{ONEROWPERIND}, \code{INDLABEL}, \code{POPID}, 
  \code{EXTRACOL}/\code{EXTRACOLS}, \code{MARKERNAME}/\code{MARKERNAMES}. If \code{RECESSIVEALLELE} or 
  \code{MAPDISTANCE} are present, they can be pluralized.

\code{read.data} uses the \code{\link[data.table]{fread}} function from \pkg{\href{https://cran.r-project.org/web/packages/data.table/index.html}{data.table}} 
  to rapidly read in the data file. The 
  loaded data file is therefore of \href{http://127.0.0.1:23982/library/base/html/class.html}{class} \code{data.table} and \code{data.frame}. 
  A \code{data.table} has slightly different syntax to a standard \code{data.frame}
  for data manipulation. Hence if you wish to manipulate the imported data file in any way, such as subsetting, and are not comfortable using 
  \code{data.table} syntax, convert the file to a standard \code{data.frame} first. However, it must be converted back to a 
  \code{data.table} before use in downstream functions.
}
\examples{

\dontrun{
#For a table with complete header row, \\code{PLOIDY=2} (default) and one column per locus:
dat=read.data("filename.ext",
nprecol=2,
MISSINGVAL=NA)

#For a structure-format file with only locus names on the header row and \\code{PLOIDY} columns per locus:
dat2=read.data(file="filename.ext",
precol.headers=0,
nprecol=2,
EXTRACOL=0,      #Same as default; included for clarity#
INDLABEL=1,      #Same as default; included for clarity#
LOCDATA=0,       #Same as default; included for clarity#
MAPDISTANCE=0,   #Same as default; included for clarity#
MARKERNAME=1,    #Same as default; included for clarity#
MISSINGVAL=-9,
NUMINDS=354,     #Not mandatory#
NUMLOCI=77,
ONEROW=1,
PHASED=0,        #Same as default; included for clarity#
PHENOTYPE=0,     #Same as default; included for clarity#
PLOIDY=2,        #Same as default; included for clarity#
POPFLAG=0,       #Same as default; included for clarity#
POPID=1,         #Same as default; included for clarity#
RECESSIVEALLELE=0#Same as default; included for clarity#
)
}
}
\author{
Richard Ian Bailey \email{richardianbailey@gmail.com}
}
