% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{rtlnorm}
\alias{rtlnorm}
\title{Internal functions not to be exported.}
\usage{
rtlnorm(n, mean = 0, sd = 1, lower = -Inf, upper = Inf)
}
\description{
Internal functions not to be exported.
}
\author{
Richard Ian Bailey \email{richardianbailey@gmail.com}
}
