% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtnorm.r, R/utils.r
\name{rtnorm}
\alias{rtnorm}
\title{Internal function that samples from a truncated normal distribution.}
\usage{
rtnorm(n, mean = 0, sd = 1, lower = -Inf, upper = Inf)

rtnorm(n, mean = 0, sd = 1, lower = -Inf, upper = Inf)
}
\description{
Internal function that samples from a truncated normal distribution.

Internal functions not to be exported.
}
\author{
Richard Ian Bailey \email{richardianbailey@gmail.com}

Richard Ian Bailey \email{richardianbailey@gmail.com}
}
