% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.prep.R
\name{data.prep}
\alias{data.prep}
\title{Prepare data for hybrid index and genomic cline estimation.}
\usage{
data.prep(
  data,
  loci,
  alleles,
  marker.info = NULL,
  S0 = NULL,
  S1 = NULL,
  sourceAbsent = FALSE,
  precols,
  INDLABEL.name = "INDLABEL",
  POPID.name = "POPID",
  max.S.MAF = 0.5,
  min.diff = 0,
  min.allele.copies.S0 = 0,
  min.allele.copies.S1 = 0,
  AF.CIoverlap = TRUE,
  return.genotype.table = FALSE,
  return.locus.table = FALSE,
  prior.prop_1 = c(1, 1)
)
}
\arguments{
\item{data}{A \code{data} object produced by \code{read.data}, or a custom \code{data.table} in the same format.}

\item{loci}{A \code{loci} object produced by \code{read.data}, or a custom character vector with locus names.}

\item{alleles}{An \code{alleles} object produced by \code{read.data}, or a custom \code{data.table} in the same format.}

\item{marker.info}{A \code{marker.info} object uploaded using \code{read.data}, or a custom \code{data.table} in the same format, 
with one row per locus and any number of columns. Default is \code{NULL}.}

\item{S0}{Character vector. Identifiers of the test subjects representing the first parental reference set (\dQuote{Source 0}). 
Allele frequencies in this set represent a hybrid index of 0. By default \code{data.prep} searches for these identifiers in the 
\code{POPID} column of the \code{data} object, but a different column can be designated using \option{POPID.name}. Default is 
\code{NULL}, and parental reference samples are not needed if parental allele frequencies are included in a \code{marker.info} 
file in the correct format.}

\item{S1}{Character vector. Identifiers of the test subjects representing the second parental reference set (\dQuote{Source 1}). 
Allele frequencies in this set represent a hybrid index of 1. By default \code{data.prep} searches for these identifiers in the 
\code{POPID} column of the \code{data} object, but a different column can be designated using \option{POPID.name}.  Default is 
\code{NULL}, and parental reference samples are not needed if parental allele frequencies are included in a \code{marker.info} 
file in the correct format.}

\item{sourceAbsent}{Logical. Whether source (parental reference) samples are absent from the dataset. If \code{TRUE}, the parental 
allele frequencies will be taken from the \code{marker.info} object uploaded using \code{read.data}. 
Default is \code{FALSE}.}

\item{precols}{Character vector. A \code{precols} object produced by \code{read.data}, or a custom character vector with 
names of all pre-marker columns in the \code{data} object. No default.}

\item{INDLABEL.name}{Character string. The name of the column in \code{data} designated as \code{INDLABEL}. 
Default is \dQuote{INDLABEL}.}

\item{POPID.name}{Character string. The name of the column in \code{data} designated as \code{POPID}. If the \code{data} object 
contains only one pre-marker column, such as \code{INDLABEL}, this must be specified as the \option{POPID.name} to avoid an 
error. The main purpose of this option is to declare which column contains the identifiers for parental reference samples, and 
any non-marker column can be used. Although it must be specified at least by the default, it is ignored if there are no parental 
reference samples in the dataset. Default is \dQuote{POPID}.}

\item{max.S.MAF}{Numeric. Locus filtering option. Removes loci for which the smaller of the two parental minor allele frequencies 
is greater than this value. Included because only one parental reference minor allele frequency needs to be close to 
zero for the locus to be informative. Default is \code{0.5} (no filtering).}

\item{min.diff}{Numeric. Locus filtering option. Removes loci for which the difference in parental reference allele frequency 
is less than this value. Default is \code{0} (no filtering).}

\item{min.allele.copies.S0}{Numeric. Locus filtering option based on sample size. Removes loci for which the number of allele 
copies in the \code{S0} parental reference set, excluding missing data, is less than this value. This option is only available 
when parental reference samples \code{S0} and \code{S1} are included in the dataset. Default is \code{0} (no filtering).}

\item{min.allele.copies.S1}{Numeric. Locus filtering option based on sample size. Removes loci for which the number of allele 
copies in the \code{S1} parental reference set, excluding missing data, is less than this value. This option is only available 
when parental reference samples \code{S0} and \code{S1} are included in the dataset. Default is \code{0} (no filtering).}

\item{AF.CIoverlap}{Logical. Recommended locus filtering option, based on posterior credible intervals of parental allele frequencies. 
If \code{FALSE}, removes loci for which there is overlap between the allele frequency 95 percent credible intervals of the 
two parental reference populations. If parental reference sets are identified, the posterior credible intervals are calculated 
internally. If this option is to be used on parental allele frequencies uploaded in a \code{marker.info} file, that file must 
contain columns entitled \sQuote{S0.prop_1_cred.int.upper} and \sQuote{S1.prop_1_cred.int.lower}, to indicate some kind of 
confidence limits for the parental allele frequencies. Only the upper limit is needed for \code{S0}, and only the lower limit is 
needed for \code{S1}. Default is \code{TRUE} (no filtering).}

\item{return.genotype.table}{Logical. Whether to create a genotype table with individual genotypes for each locus scored from 
\code{0} to \code{PLOIDY}, based on their number of copies of the designated \code{S1} allele (the allele with relatively higher 
frequency in the \code{S1} parental reference set). Default is \code{FALSE}.}

\item{return.locus.table}{Logical. Whether to return a table with one row per locus containing a wider array of locus-level data, 
such as parental allele frequency credible intervals and number of non-missing allele copies in each parental reference set. 
Default is \code{FALSE} to save memory, but this could be useful in many circumstances.}

\item{prior.prop_1}{Numeric vector. Values of \code{shape1} and \code{shape2} parameters for a beta-distributed prior
for parental allele frequencies (proportions) of the \code{S1_allele}. Default is \code{shape1=shape2=1} (uniform distribution).}
}
\value{
list containing at least two components: (1) \code{data.prep}, a \code{data.table} and \code{data.frame} long-form version of 
  the imported data table, with one row per non-missing allele copy and the data columns required for analysis, listed below; (2) 
  \code{Nloci.postfilter}, a numeric scalar indicating the number of loci remaining after filtering. If parental reference samples are 
  identified, a third output is returned, \code{sourceInfo}, a \code{data.table} and \code{data.frame} listing all the unique entries in 
  the declared \sQuote{POPID.name} column (default \sQuote{POPID}) and indicating whether they represent \sQuote{S0}, \sQuote{S1} or 
  \sQuote{TEST}. The two optional outputs are: (1) \code{geno.data}, a \code{data.table} and \code{data.frame} containing genotypes for 
  each locus scored from \code{0} to \code{PLOIDY}, based on their number of copies of the designated \code{S1} allele (the allele with 
  relatively higher frequency in the \code{S1} parental reference set); (2) \code{locus.data}, a \code{data.table} and \code{data.frame} 
  with one row per locus and containing useful locus-specific information, including the contents of any uploaded \code{marker.info} 
  object.

  \code{data.prep} contains the pre-marker columns and the following fields, plus any columns from a declared \code{marker.info} object:
  \item{Source}{whether the sample is from parental reference 0 (\sQuote{S0}), parental reference 1 (\sQuote{S1}) or is a test individual 
     (\sQuote{TEST}).}
  \item{locus}{the locus names.}
  \item{S0.prop_1}{frequency (proportion) of the \code{S1_allele} in \code{S0}.}
  \item{S1.prop_1}{frequency (proportion) of the \code{S1_allele} in \code{S1}.}
  \item{Source_allele}{numeric. Whether the allele is the \code{S0_allele} (\sQuote{0}) or \code{S1_allele} (\sQuote{1}).}

  \code{locus.data} additionally contains, at a minimum:
  \item{S0_allele}{name of the allele with relatively higher frequency in \code{S0}.}
  \item{S1_allele}{name of the allele with relatively higher frequency in \code{S1}.}
  \item{Source.afdiff}{difference in allele frequency between \code{S0} and \code{S1}.}
  \item{S0.MAF}{\code{S0} minor allele frequency.}
  \item{S1.MAF}{\code{S1} minor allele frequency.}
  \item{min.MAF}{smaller of \code{S0.MAF} and \code{S1.MAF}.}

  If parental reference samples are present, \code{locus.data} further contains:
  \item{S0.N_1}{number of non-missing copies of the \code{S1_allele} in \code{S0}.}
  \item{S1.N_1}{number of non-missing copies of the \code{S1_allele} in \code{S1}.}
  \item{S0.prop_0}{frequency (proportion) of the \code{S0_allele} in \code{S0}.}
  \item{S1.prop_0}{frequency (proportion) of the \code{S0_allele} in \code{S1}.}
  \item{N_allele_copies.S0}{total number of non-missing allele copies in \code{S0}.}
  \item{N_allele_copies.S1}{total number of non-missing allele copies in \code{S1}.}
  \item{S0.prop_1_shape1}{shape1 parameter estimate for the posterior beta distribution of the \code{S1_allele} frequency in \code{S0}.}
  \item{S0.prop_1_shape2}{shape2 parameter estimate for the posterior beta distribution of the \code{S1_allele} frequency in \code{S0}.}
  \item{S0.prop_1_postmode}{mode of the beta-distributed posterior for the frequency of the \code{S1_allele} in \code{S0}.}
  \item{S0.prop_1_postmean}{mean of the beta-distributed posterior for the frequency of the \code{S1_allele} in \code{S0}.}
  \item{S0.prop_1_cred.int.lower}{lower 95 percent credible interval (2.5 percent quantile of the posterior beta distribution) of the frequency of the \code{S1_allele} in \code{S0}.}
  \item{S0.prop_1_cred.int.upper}{upper 95 percent credible interval (97.5 percent quantile of the posterior beta distribution) of the frequency of the \code{S1_allele} in \code{S0}.}
  \item{S1.prop_1_shape1}{shape1 parameter estimate for the posterior beta distribution of the \code{S1_allele} frequency in \code{S1}.}
  \item{S1.prop_1_shape2}{shape2 parameter estimate for the posterior beta distribution of the \code{S1_allele} frequency in \code{S1}.}
  \item{S1.prop_1_postmode}{mode of the beta-distributed posterior for the frequency of the \code{S1_allele} in \code{S1}.}
  \item{S1.prop_1_postmean}{mean of the beta-distributed posterior for the frequency of the \code{S1_allele} in \code{S1}.}
  \item{S1.prop_1_cred.int.lower}{lower 95 percent credible interval (2.5 percent quantile of the posterior beta distribution) of the frequency of the \code{S1_allele} in \code{S1}.}
  \item{S1.prop_1_cred.int.upper}{upper 95 percent credible interval (97.5 percent quantile of the posterior beta distribution) of the frequency of the \code{S1_allele} in \code{S1}.}
  \item{S1.prop_0_shape1}{shape1 parameter estimate for the posterior beta distribution of the \code{S0_allele} frequency in \code{S1}.}
  \item{S1.prop_0_shape2}{shape2 parameter estimate for the posterior beta distribution of the \code{S0_allele} frequency in \code{S1}.}
  \item{S0.prop_0_shape1}{shape1 parameter estimate for the posterior beta distribution of the \code{S0_allele} frequency in \code{S0}.}
  \item{S0.prop_0_shape2}{shape2 parameter estimate for the posterior beta distribution of the \code{S0_allele} frequency in \code{S0}.}
}
\description{
Prepare data for hybrid index and genomic cline estimation.
}
\details{
The primary output of \code{data.prep} is a long-form \code{data.table} with 1 row per non-missing allele copy in the 
  data set, to be used in downstream hybrid index and genomic cline estimation. It also optionally produces (1) a locus table,
  with one row per locus and containing useful locus-level data (see \sQuote{Value}), and (2) a genotype table, with genotypes 
  scored according to the number of alleles from parental reference set \code{S1}. This may be useful for example in estimating 
  admixture-induced (\sQuote{ancestry}) linkage disequilibria. \code{data.prep} uses objects produced by \code{read.data} as input, 
  or custom objects in the same format.

Even when all filtering options are left at default values, loci with exactly zero allele frequency difference between S0 and S1 will 
  be removed. Therefore, it is often likely that the resulting analysis table will contain fewer loci than were loaded by \code{read.data}.

The optional \code{marker.info} file should be a file with one row per marker and any number of columns (including the marker name in 
  a column named \sQuote{locus}), each for a variable of interest for grouping markers, such as chromosome, map location, gene 
  annotations, sliding window identifiers etc. These variables can then be used downstream for statistical comparison of hybrid index 
  or clines among groups of loci. In the absence of parental reference samples, certain column names and information must be included 
  in the \code{marker.info} file. See examples here and in \code{read.data}.

The locus data object includes posterior mean, mode and upper and lower 95% credible intervals for frequency of the Source 1 allele
  for each locus, assuming beta-distributed proportions. Note that if the locus is fixed in a parental reference set, its posterior 
  mode will be 0 or 1 and will fall outside the credible intervals due to the influence of the non-zero prior shape parameters. 
  If prior shape parameters are set to zero (equivalent to no prior), the best posterior estimate is the mean rather than the mode, 
  but in this case credible intervals cannot be estimated when the locus is fixed for one allele.
}
\examples{

\dontrun{
###############################################################
#Read in and prepare a file with no parental reference samples#
###############################################################

#Create an input data file and save to the working directory.
cat("INDLABEL POPID chr1:001 chr1:002","ind1 pop1 A A","ind1 pop1 A B","ind2 pop1 B B","ind2 pop1 B B","ind3 pop2 NA A","ind3 pop2 B A", file = "ex.data", sep = "\n")

#Create a marker.info file with the necessary parental reference information and save to the working directory.
cat("locus refAllele alternateAllele S0.prop_r S1.prop_r type","chr1:001 A B 0.1 0.9 intronic","chr1:002 B A 0.8 0.2 exonic", file = "ex_marker_info2.data", sep = "\n")

#Read in the data and marker.info file.
dat <- read.data(file="ex.data",nprecol=2,NUMINDS=3,NUMLOCI=2,MISSINGVAL=NA,marker.info.file = "ex_marker_info2.data", sourceAbsent = TRUE)

#Run data.prep.
prepdata=data.prep(
 data=dat$data,               #part of the read.data output object#
 loci=dat$loci,               #part of the read.data output object#
 sourceAbsent = TRUE,         #Must be set to TRUE in the absence of parental reference samples. Default is FALSE#
 marker.info=dat$marker.info, #must be specified if sourceAbsent = TRUE; make sure it contains the headers and their 
                              #contents as specified above when 'sourceAbsent=TRUE'#
 alleles=dat$alleles,         #part of the read.data output object#
 #S0,                         #character vector identifying parental reference S0 samples. Leave blank in the absence of parental reference samples#
 #S1,                         #character vector identifying parental reference S1 samples. Leave blank in the absence of parental reference samples#
 precols=dat$precols,         #part of the read.data output object#
 return.genotype.table=TRUE,  #This returns an table, with loci in columns, of diploid (or other ploidy) genotypes, coded as number of copies of the 
                              #allele with higher frequency in S1 than S0#
 return.locus.table=TRUE      #Returns a table with one row per locus and all the individual marker data, including data already uploaded plus values 
                              #calculated within this function#
)

#####################################################################
#Read in and prepare a file that includes parental reference samples#
#####################################################################

#Make a new data input file with more entries, including S0 and S1 samples, in plink structure format but with missing data recoded to NA. This includes 
#a MAPDISTANCE row below the marker names.
cat("chr1:001 chr1:002","-3 3 2 4",
 "ind1 pop1 A A A B","ind2 pop1 B B B B","ind3 pop2 NA B A A", 
 "ind4 pop3 A A A A","ind5 pop3 A B A A","ind6 pop3 A A A A",
 "ind7 pop3 NA A A A","ind8 pop4 A NA B B","ind9 pop4 B B B B",
 "ind10 pop4 B B B B","ind11 pop5 A A B B",
 file = "ex3.data", sep = "\n")

#pop3 represents S0; pop4 and pop5 represent S1.

#Run read.data and data.prep.
dat3 <- read.data(file="ex3.data",MISSINGVAL=NA,NUMINDS=11,nprecol=2,ONEROW=1,MAPDISTANCE=1,precol.headers=0)

prepdata= data.prep(
 data=dat3$data,
 loci=dat3$loci,
 sourceAbsent = FALSE,                         #default#
 #marker.info=dat$marker.info,                 #a marker.info file is not obligatory when parental reference samples are declared#
 alleles=dat3$alleles,
 S0="pop3",                                    #first parental reference set; must be specified when sourceAbsent = FALSE#
 S1=c("pop4","pop5"),                          #second parental reference set; must be specified when sourceAbsent = FALSE#
 precols=dat3$precols,
 return.genotype.table=TRUE,                   #default is FALSE to save memory#
 return.locus.table=TRUE                       #default is FALSE to save memory#
)

#########################################
#Now completely exclude the POPID column#
#########################################

#Just to show that the functions work in the absence of this column#

#First with no parental reference samples. Modify the plink format 'ex3.data' above to remove the POPID column, 
#and use the existing marker info file, 'ex_marker_info2.data'#

cat("chr1:001 chr1:002", "-3 3 2 4",
 "ind1 A A A B","ind2 B B B B","ind3 NA B A A", 
 "ind4 A A A A","ind5 A B A A","ind6 A A A A",
 "ind7 NA A A A","ind8 A NA B B","ind9 B B B B","ind10 B B B B","ind11 A A B B",
 file = "ex4.data", sep = "\n")

#you must specify 'POPID=0' (meaning the column is absent) in read.data.
dat4 <- read.data(file="ex4.data",POPID=0,nprecol=1,NUMINDS=11,NUMLOCI=2,MISSINGVAL=NA,marker.info.file="ex_marker_info2.data",sourceAbsent=TRUE,ONEROW=1,MAPDISTANCE=1,precol.headers=0)

#Now prepare the data for analysis, this time we're not creating geno.data or locus.data objects.
prepdata= data.prep(
 data=dat4$data,
 loci=dat4$loci,
 sourceAbsent = TRUE,         #default is FALSE#
 marker.info=dat4$marker.info,#must be specified if sourceAbsent = TRUE#
 alleles=dat4$alleles,
 #S0,                         #leave blank in the absence of parental reference samples#
 #S1,                         #leave blank in the absence of parental reference samples#
 POPID.name="INDLABEL",     #***must be specified if there's no POPID column, to avoid an error***#
 precols=dat4$precols
)

#Now including parental reference samples, again with no POPID column.

#Create and save a marker info file. Not obligatory in the presence of parental reference samples.
cat("locus type","chr1:001 intronic","chr1:002 exonic", file = "ex_marker_info.data", sep = "\n")

#Read data in and set sourceAbsent=FALSE (the default setting) so that the code doesn't search for the obligatory columns and throw up an error.
dat5 <- read.data(file="ex4.data",POPID=0,nprecol=1,NUMINDS=11,NUMLOCI=2,MISSINGVAL=NA,marker.info.file="ex_marker_info.data",
 sourceAbsent=FALSE,ONEROW=1,MAPDISTANCE=1,precol.headers=0)

#Prepare the analysis object.
prepdata= data.prep(
 data=dat5$data,
 loci=dat5$loci,
 sourceAbsent = FALSE,                               #default is FALSE#
 marker.info=dat5$marker.info,                       #this now contains only the locus (obligatory) and type columns#
 alleles=dat5$alleles,
 S0=c("ind4","ind5","ind6","ind7"),                  #must be specified when sourceAbsent = FALSE#
 S1=c("ind8","ind9","ind10","ind11"),                #must be specified when sourceAbsent = FALSE#
 POPID.name = "INDLABEL",                            #***must be specified if there's no POPID column; the S0 and S1 sample 
                                                     #references must be present in the column specified here***#
 precols=dat5$precols,
 return.genotype.table=TRUE,                         #Optional#
 return.locus.table=TRUE                             #Optional#
)

#################################################################################
#Locus filtering based on parental reference allele frequencies and sample sizes#
#################################################################################

#Upload a real dataset (Italian sparrows), included as part of the package.
dat=read.data("RB_Italy_ggcline_precol_headers_haploidND2.csv",nprecol=2,MISSINGVAL=NA,NUMINDS=569)

#The file contains data for 77 diploid markers for 569 male sparrow individuals.
#Run data.prep including the different filtering options. Try hashing out or changing these options to see how it affects the number of 
#loci in the resulting analysis table (see prepdata$Nloci.postfilter in the data.prep object).
prepdata=data.prep(
 data=dat$data,
 loci=dat$loci,
 alleles=dat$alleles,
 S0=c("Kralove","Oslo"), #POPID names for the first parental reference set#
 S1="LesinaSPANISH",     #POPID names for the second parental reference set#
 precols=dat$precols,
###Filtering below###
 max.S.MAF = 0.2,               #minor allele frq must be below this value in at least one of S0 or S1#
 min.diff = 0.2,                #loci with parental allele frequency difference < this value will be removed#
 min.allele.copies.S0 = 30,     #loci with fewer non-missing allele copies in the S0 parental reference set will be removed#
 min.allele.copies.S1 = 12,     #in this dataset the S1 sample size is much smaller than S0, so I'm being less strict with filtering for S1 sample size#
 AF.CIoverlap = FALSE,          #***RECOMMENDED*** filtering option if parental reference samples are included - removes all loci for which there is 
                                #overlap between S0 and S1 in the Bayesian posterior 95\% credible intervals of allele frequency# 
)

unlink("ex_marker_info.data")#Tidy up#
unlink("ex_marker_info2.data")#Tidy up#
unlink("ex.data")#Tidy up#
unlink("ex3.data")#Tidy up#
unlink("ex4.data")#Tidy up#
}
}
\author{
Richard Ian Bailey \email{richardianbailey@gmail.com}
}
