% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcline.r
\name{ggcline}
\alias{ggcline}
\title{Estimate genomic cline parameters, credible intervals and p values.}
\usage{
ggcline(data.prep.object, esth.object, test.subject = "locus",
  poolv = FALSE, poolcentre = FALSE, include.Source = FALSE,
  return.likmeans = FALSE, read.data.precols, fix.subject.v = FALSE,
  fix.value.v, fix.subject.centre = FALSE, fix.value.centre,
  plot.test.subject = NULL, plot.col = NULL, plot.ylim = c(-5, 5),
  plot.pch.v.centre = c(1, 3), prior.centre = c(0, sqrt(50)),
  prior.logv = c(0, sqrt(10)), nitt, burnin, start.v = NULL,
  start.centre = NULL, init.var.v = NULL, init.var.centre = NULL,
  init.cov.vcentre = NULL, print.k = 50)
}
\arguments{
\item{data.prep.object}{Name of the \code{data.prep} object produced by the \code{data.prep} function.}

\item{esth.object}{Name of the \code{esth} object.}

\item{test.subject}{Character string. Name of the field identifying subjects for genomic cline estimation. Default \dQuote{locus}.}

\item{poolv}{Logical. Whether to pool the rate parameter \code{v} across all test subjects. Default \code{FALSE}.}

\item{poolcentre}{Logical. Whether to pool the \code{centre} parameter across all test subjects. Default \code{FALSE}.}

\item{include.Source}{Logical. Whether to include the designated parental reference sets as test subjects in genomic cline 
estimation. Default \code{FALSE}.}

\item{return.likmeans}{Logical. Whether to return a table of likelihood calculations. Required for \code{compare.models}.}

\item{read.data.precols}{A \code{precols} object produced by \code{read.data}, or a custom character vector with 
names of all pre-marker columns in the \code{data}.}

\item{fix.subject.v}{Logical default, otherwise a character vector. List of test subjects (normally locus names) 
for which you wish to fix rather than estimate the rate parameter \code{v}. Used for model 
comparison with \code{compare.models}.}

\item{fix.value.v}{Logical default, otherwise a numeric vector of length 1 or the same length as \code{fix.subject.v}. 
The fixed \code{v} value(s) for the test subject(s) listed in \code{fix.subject.v}.}

\item{fix.subject.centre}{Logical default, otherwise a character vector. List of test subjects (normally locus names) 
for which you wish to fix rather than estimate the \code{centre} parameter . Used for model 
comparison with \code{compare.models}.}

\item{fix.value.centre}{Logical default, otherwise a numeric vector of length 1 or the same length as \code{fix.subject.centre}. 
The fixed \code{centre} value(s) for the test subject(s) listed in \code{fix.subject.centre}.}

\item{plot.test.subject}{Character vector. List of up to 3 test subjects for which you wish to plot the posterior distribution values in real time. 
Rate parameter \code{v} is plotted with open circles, and \code{centre} with plus signs, both the same colour for the same 
test subject. Default \code{NULL}.}

\item{plot.col}{Character vector. List of up to 3 colours to plot the \code{plot.test.subject} test subjects. Default \code{NULL}.}

\item{plot.ylim}{Numeric vector. Upper and lower plot limits of the y axis. Default \code{c(-5,5)}.}

\item{plot.pch.v.centre}{Numeric vector. Size of data points for the two parameters. Default \code{c(1,3)}.}

\item{prior.centre}{Numeric vector. Mean and sd of the normally distributed prior for \code{centre}. Default \code{c(0,sqrt(50))}.}

\item{prior.logv}{Numeric vector. Mean and sd of the normally distributed prior for \code{v}. Default \code{c(0,sqrt(50))}.}

\item{nitt}{The total number of MCMC iterations including burnin. At least 10000 is recommended, more if one parameter is pooled.}

\item{burnin}{Numeric. The number of burnin MCMC iterations. At least 5000 is recommended, more if one parameter is pooled.}

\item{start.v}{Numeric vector. Optional starting \code{v} values for the MCMC, which are otherwise sampled randomly. Default \code{NULL}.}

\item{start.centre}{Numeric vector. Optional starting \code{centre} values for the MCMC, which are otherwise sampled randomly. Default \code{NULL}.}

\item{init.var.v}{Numeric. Optional starting variance of the \code{v} parameter proposal distribution. Internal value works 
well on tested data sets. Default \code{NULL}.}

\item{init.var.centre}{Optional starting variance of the \code{centre} parameter proposal distribution. Internal value works 
well on tested data sets. Default \code{NULL}.}

\item{init.cov.vcentre}{Optional starting covariance of the \code{v} and \code{centre} multivariate normal proposal distribution. 
Internal value (0) works well on tested data sets.
well on tested data sets. Default \code{NULL}.}

\item{print.k}{The iteration is printed to screen on multiples of this number. Default is \code{50}.}
}
\value{
list with two mandatory components, \code{gc}, a \code{data.table} and \code{data.frame} with the estimated 
  genomic cline parameters and their credible intervals, and \code{test.subject}. The optional output \code{likmeans} 
  is needed downstream 
  when model comparison is carried out with \code{compare.models}.

  \code{gc} contains the \code{test.subject} and following fields:
  \item{v_mean}{the rate parameter estimate.}
  \item{v_lower_95}{rate parameter lower 95 percent credible interval (inverse-logged 2.5 percent quantile of the posterior normal distribution for log(v)).}
  \item{v_upper_95}{rate parameter upper 95 percent credible interval (inverse-logged 97.5 percent quantile of the posterior normal distribution for log(v)).}
  \item{v_pvalue}{p value for the rate parameter. Quantile of the null value (log(1)) given the posterior normal distribution for log(v).}
  \item{centre_mean}{mean of the posterior normal distribution.}
  \item{centre_lower_95}{lower 95 percent credible interval (2.5 percent quantile of the posterior normal).}
  \item{centre_upper_95}{upper 95 percent credible interval (97.5 percent quantile of the posterior normal).}
  \item{centre_pvalue}{p value. Quantile of the null value (0.5) given the posterior normal distribution for centre.}
  \item{npar}{number of parameters.}
  \item{sum_npar}{sum number of parameters across all samples.}
}
\description{
Estimate genomic cline parameters, credible intervals and p values.
}
\details{
\code{ggcline} assumes the posterior of \code{log(v)} is normally distributed, and the results for this parameter are 
  the inverse logs of the estimated mean and upper and lower quantiles of \code{log(v)}.

The null value \code{v} = 1, and for \code{centre} = 0.5. If both parameters are fixed to these or other values, only \code{nitt=2} 
  and \code{burnin=0} are required.
}
\examples{

\dontrun{
gc1=ggcline(
data.prep.object=prepdata$data.prep,
esth.object=hindlabel,
poolv=FALSE,poolcentre=FALSE,	#No pooling of parameters for this run#
test.subject="locusgroup",		#Use a field other than the default#
return.likmeans=TRUE,
read.data.precols=dat$precols,
fix.subject.v=c("locusgroup1","locusgroup2"),fix.value.v=1,	#Fixing a subset of test subjects for v#
fix.subject.centre=c("locusgroup1","locusgroup2"),fix.value.centre=0.5,	#Fixing a subset of test subjects for centre#
plot.test.subject=c("locusgroup1"),	#Real-time plot of the two parameters for locusgroup1#
plot.col=c("orange"),
plot.ylim=c(-2,3),
nitt = 10000, burnin = 5000, print.k = 50)
}
}
\author{
Richard Ian Bailey \email{richardianbailey@gmail.com}
}
