% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clinecurve.r
\name{plot_clinecurve}
\alias{plot_clinecurve}
\title{Plot the fitted cline for one or more test subjects.}
\usage{
plot_clinecurve(ggcline.object, data.prep.object, esth.object,
  test.subject = "INDLABEL", include.Source = FALSE, cline.locus,
  locus.column = "locus", cline.col = "black", null.line.locus = NULL,
  null.line.col = "black", plot.data = NULL, data.col = NULL,
  plot.genotype = FALSE, PLOIDY, cline.centre.line = NULL,
  cline.centre.col = "black", ...)
}
\arguments{
\item{ggcline.object}{A \code{gc} object from \code{ggcline}.}

\item{data.prep.object}{A \code{data.prep} object from \code{data.prep}.}

\item{esth.object}{An \code{hi} object from \code{esth}.}

\item{test.subject}{Character string. The test subject used in \code{esth}.}

\item{include.Source}{Logical. If plotting individual data points, whether to include parental reference data. Default \code{FALSE}. 
Set to \code{TRUE} if parental reference sets were included in \code{ggcline}.}

\item{cline.locus}{Character vector. Names of the loci or \code{ggcline} test subjects for which to plot cline curves.}

\item{locus.column}{Character string. Name of the field containing the test subjects. Default \dQuote{locus}.}

\item{cline.col}{Character vector. Colours for the plotted cline curves. Default \dQuote{black} (repeated to the number of loci plotted).}

\item{null.line.locus}{Character vector. Names of loci for which to plot a black dashed line showing the null expectation. Default \code{NULL}.}

\item{null.line.col}{Character vector. Colours for the plotted null dashed lines. Default \dQuote{black} (repeated to the number of loci plotted).}

\item{plot.data}{Logical default, otherwise a character vector. Names of loci for which to plot data points on the figure.}

\item{data.col}{Character vector. Colours for the plotted data (one for each locus/\code{ggcline} test subject).}

\item{plot.genotype}{Logical. Whether to plot genotypes of individuals (or \code{esth} test subjects) scaled from 0 to 1 rather than individual 
allele copies (which will all take values 0 or 1).}

\item{PLOIDY}{Numeric. The number of allele copies for each \code{esth} individual test subject (typically \code{INDLABEL}).}

\item{cline.centre.line}{Character vector. Names of \code{ggcline} test subjects (loci) for which to plot a vertical dashed line 
indicating the cline centre. Default \code{NULL}.}

\item{cline.centre.col}{Character vector. Colours for the plotted centre lines.}

\item{...}{Further graphical parameters.}
}
\value{
a plot.
}
\description{
Plot the fitted cline for one or more test subjects.
}
\examples{

\dontrun{
plot_clinecurve(
ggcline.object=gc1$gc,
data.prep.object=prepdata$data.prep,
esth.object=hindlabel$hi,
cline.locus="locusgroup1",                                                                  #Plotting the cline for only one locus#
locus.column="locusgroup",
#cline.col=c("orange","blue","green","red","magenta"),                                      #If you had chosen to plot 5 loci#
cline.col="green",
#null.line.locus=c("locusgroup1","locusgroup2","locusgroup3","locusgroup4","locusgroup5"),  #If you had chosen to plot 5 loci#
null.line.locus="locusgroup1",
#null.line.col=c("orange","blue","green","red","magenta"),                                  #If you had chosen to plot 5 loci#
null.line.col="green",
plot.data="locusgroup1",
data.col="black",
plot.genotype=TRUE,
PLOIDY=2,
cline.centre.line="locusgroup1",
cline.centre.col="green"
#cline.centre.line=c("locusgroup1","locusgroup2","locusgroup3","locusgroup4","locusgroup5"),#If you had chosen to plot 5 loci#
#cline.centre.col=c("orange","blue","green","red","magenta")                                #If you had chosen to plot 5 loci#
)
#Add a title:
title(main = "locusgroup1",xlab="Hybrid index",
ylab="Genotype frequency",cex.main=1.5,cex.lab=1.5)
}
}
\author{
Richard Ian Bailey \email{richardianbailey@gmail.com}
}
